static/tests/samesite/fetch.html

<!DOCTYPE html>
<script src="/tests/resources/testharness.js"></script>
<script src="/tests/resources/testharnessreport.js"></script>
<script src="/tests/resources/cookie-helper.js"></script>
<script>
  var HTTP_COOKIE = "cookie_via_http";

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Lax")
      .then(_ => {
        return assert_http_cookie(ORIGIN, HTTP_COOKIE, COOKIE_VALUE, true);
      });
  }, "Same-host requests are laxly same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Strict")
      .then(_ => {
        return assert_http_cookie(ORIGIN, HTTP_COOKIE, COOKIE_VALUE, true);
      });
  }, "Same-host requests are strictly same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(SUBDOMAIN_ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Lax")
      .then(_ => {
        return assert_http_cookie(SUBDOMAIN_ORIGIN, HTTP_COOKIE, COOKIE_VALUE, true);
      });
  }, "Requests to subdomains are laxly same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(SUBDOMAIN_ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Strict")
      .then(_ => {
        return assert_http_cookie(SUBDOMAIN_ORIGIN, HTTP_COOKIE, COOKIE_VALUE, true);
      });
  }, "Requests to subdomains are strictly same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(CROSS_SITE_ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Strict")
      .then(_ => {
        return assert_http_cookie(CROSS_SITE_ORIGIN, HTTP_COOKIE, COOKIE_VALUE, false);
      });
  }, "Cross-origin requests are not same-site.");
</script>
